Se actualiza con la version 1.3.2 de desarrollo

+++ src/cl/buildersoft/web/servlet/timectrl/machine/DeleteEmployee.java
@@ -10,8 +10,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import cl.buildersoft.framework.database.BSBeanUtils;
 import cl.buildersoft.framework.util.BSConnectionFactory;
-import cl.buildersoft.framework.util.BSHttpServlet;
-import cl.buildersoft.timectrl.api._zkemProxy;
+import cl.buildersoft.framework.web.servlet.BSHttpServlet_;
+import cl.buildersoft.timectrl.api.com4j._zkemProxy;
 import cl.buildersoft.timectrl.business.beans.Machine;
 import cl.buildersoft.timectrl.business.services.MachineService2;
 import cl.buildersoft.timectrl.business.services.impl.MachineServiceImpl2;
@@ -20,7 +20,7 @@ import cl.buildersoft.timectrl.business.services.impl.MachineServiceImpl2;
  * Servlet implementation class DeleteEmployee
  */
 @WebServlet("/servlet/timectrl/machine/DeleteEmployee")
-public class DeleteEmployee extends BSHttpServlet {
+public class DeleteEmployee extends BSHttpServlet_ {
 	private static final long serialVersionUID = 8751209270314151392L;
 
 	protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -39,7 +39,9 @@ public class DeleteEmployee extends BSHttpServlet {
 		_zkemProxy api = machineService.connect(conn, machine);
 		cf.closeConnection(conn);
 
-		machineService.deleteEmployees(api, keys);
+		if (keys != null) {
+			machineService.deleteEmployees(api, keys);
+		}
 		machineService.disconnect(api);
 		forward(request, response, "/servlet/timectrl/machine/MachineManager");
